Before (https://github.com/pivovarit/AoC_2023_go/pull/4):

```
trebuchetPart1() result: 55386, execution time: 45.125µs
trebuchetPart2() result: 54824, execution time: 9.765458ms
```

Then (https://github.com/pivovarit/AoC_2023_go/pull/22):

```
trebuchetPart1() result: 55386, execution time: 42.25µs
trebuchetPart2() result: 54824, execution time: 831.584µs
```

Then (https://github.com/pivovarit/AoC_2023_go/pull/23)

```
trebuchetPart1() result: 55386, execution time: 31.542µs
trebuchetPart2() result: 54824, execution time: 269.416µs
```

Now: 

```
trebuchetPart1() result: 55386, execution time: 31.625µs
trebuchetPart2() result: 54824, execution time: 86.333µs
```

bench results:
```
BenchmarkTrebuchetPart2-10 21637 53597 ns/op
```